fix: Move Escrow invalid bounds test out of test_fuzzing.py
- Moved `test_task_award_escrow_invalid` from `tests/test_fuzzing.py` to `tests/domains/test_workflow_auctions.py` to respect parallel execution isolation mandates.
- Changed exception catch from `ValueError` to `pydantic.ValidationError` since Pydantic V2 raises `ValidationError` for `@model_validator` failures.
- Fixed `ruff format` failures.
